refactor(tanstack): rename mutation-options to mutation and query-key to query-key-serialize
The -options suffix was redundant: files already export carAddMutationOptions,
serviceLogRemoveMutationOptions, and so on. Applied across all four modules
with a mutation directory. query-key.ts holds a single serialize function,
not keys, so its name follows what it exports.

Path-only change: no exported symbol renamed, no logic moved.
