+++ src/module/car/ownership/presentation/hooks/use-owner-profiles-for-car.test.tsx
@@ -7,7 +7,7 @@ import { ownershipDataSource } from '@/car/ownership/dependency/data-source';
 import { useOwnerProfilesForCar } from '@/car/ownership/presentation/hooks/use-owner-profiles-for-car';
 import { queryKeys } from '@/car/ownership/presentation/tanstack/query/keys';
 import { Result } from '@/common/application/result';
-import { queryKeySerialize } from '@/common/presentation/tanstack/query-key';
+import { queryKeySerialize } from '@/common/presentation/tanstack/query-key-serialize';
 import { buildUserDto } from '@/user/application/dto/user.builder';
 import { userDataSource } from '@/user/dependency/data-source';
