+++ src/module/car/ownership/presentation/hooks/use-owner-profiles-for-car.ts
@@ -4,7 +4,7 @@ import { useEffect } from 'react';
 import { queryKeys } from '@/car/ownership/presentation/tanstack/query/keys';
 import { getOwnershipsByCarIdQueryOptions } from '@/car/ownership/presentation/tanstack/query/options';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
-import { queryKeySerialize } from '@/common/presentation/tanstack/query-key';
+import { queryKeySerialize } from '@/common/presentation/tanstack/query-key-serialize';
 import { getUserByIdQueryOptions } from '@/user/presentation/tanstack/query/options';
 
 export function useOwnerProfilesForCar(carId: string) {
