+++ src/module/car/ownership/presentation/tanstack/mutation/remove.ts
