+++ src/module/car/ownership/presentation/ui/sections/total-ownerships/use-total-ownerships.test.tsx
@@ -7,7 +7,7 @@ import { ownershipDataSource } from '@/car/ownership/dependency/data-source';
 import { queryKeys } from '@/car/ownership/presentation/tanstack/query/keys';
 import { useTotalOwnershipsSection } from '@/car/ownership/presentation/ui/sections/total-ownerships/use-total-ownerships';
 import { Result } from '@/common/application/result';
-import { queryKeySerialize } from '@/common/presentation/tanstack/query-key';
+import { queryKeySerialize } from '@/common/presentation/tanstack/query-key-serialize';
 import { buildUserDto } from '@/user/application/dto/user.builder';
 import { userDataSource } from '@/user/dependency/data-source';
 import { getSessionUserQueryOptions } from '@/user/presentation/tanstack/query/options';
