+++ src/module/car/ownership/presentation/ui/sections/total-ownerships/use-total-ownerships.ts
@@ -4,7 +4,7 @@ import { useEffect } from 'react';
 import { queryKeys } from '@/car/ownership/presentation/tanstack/query/keys';
 import { getOwnershipsByOwnerIdQueryOptions } from '@/car/ownership/presentation/tanstack/query/options';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
-import { queryKeySerialize } from '@/common/presentation/tanstack/query-key';
+import { queryKeySerialize } from '@/common/presentation/tanstack/query-key-serialize';
 import { useSessionUser } from '@/user/presentation/hooks/use-session-user';
 
 export function useTotalOwnershipsSection() {
