+++ src/module/car/presentation/tanstack/mutation/add.ts
@@ -9,7 +9,7 @@ import {
   type CarsInfiniteQueryData,
   deepCopyCarsInfiniteQueryData,
   patchCarInInfiniteQueryData,
-} from '@/car/presentation/tanstack/mutation-options/shared/infinite-query-data';
+} from '@/car/presentation/tanstack/mutation/shared/infinite-query-data';
 import { queryKeys } from '@/car/presentation/tanstack/query/keys';
 
 export type CarAddMutationVariables = AddCarApiRequest & {
