+++ src/module/car/presentation/tanstack/mutation/shared/infinite-query-data.ts
