+++ src/module/car/presentation/ui/forms/edit/use-edit.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 
 import type { CarFormData } from '@/car/interface/ui/car-form.schema';
-import { carEditMutationOptions } from '@/car/presentation/tanstack/mutation-options/edit';
+import { carEditMutationOptions } from '@/car/presentation/tanstack/mutation/edit';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
 
 interface UseEditFormParams {
