+++ src/module/car/presentation/ui/sections/delete/controls/use-controls.ts
@@ -4,7 +4,7 @@ import { useRouter } from 'next/navigation';
 import { useRef } from 'react';
 
 import type { CarDto } from '@/car/application/dto/car';
-import { carRemoveMutationOptions } from '@/car/presentation/tanstack/mutation-options/remove';
+import { carRemoveMutationOptions } from '@/car/presentation/tanstack/mutation/remove';
 import { queryKeys } from '@/car/presentation/tanstack/query/keys';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
 import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
