+++ src/module/car/presentation/ui/sections/details/use-details.ts
@@ -6,7 +6,7 @@ import { getOwnershipsByCarIdQueryOptions } from '@/car/ownership/presentation/t
 import { queryKeys } from '@/car/presentation/tanstack/query/keys';
 import { getCarByIdQueryOptions } from '@/car/presentation/tanstack/query/options';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
-import { queryKeySerialize } from '@/common/presentation/tanstack/query-key';
+import { queryKeySerialize } from '@/common/presentation/tanstack/query-key-serialize';
 import { useSessionUser } from '@/user/presentation/hooks/use-session-user';
 
 type UseDetailsSectionParams = {
