+++ src/module/car/service-log/presentation/tanstack/mutation/remove.ts
