+++ src/module/car/service-log/presentation/ui/forms/edit/use-edit.ts
@@ -2,7 +2,7 @@ import { useMutation } from '@tanstack/react-query';
 
 import type { ServiceLogDto } from '@/car/service-log/application/dto/service-log';
 import type { ServiceLogFormValues } from '@/car/service-log/interface/ui/service-log-form.schema';
-import { serviceLogEditMutationOptions } from '@/car/service-log/presentation/tanstack/mutation-options/edit';
+import { serviceLogEditMutationOptions } from '@/car/service-log/presentation/tanstack/mutation/edit';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
 
 interface UseEditFormParams {
