+++ src/module/car/service-log/presentation/ui/sections/costs/use-costs.ts
@@ -4,7 +4,7 @@ import { useEffect } from 'react';
 import { queryKeys } from '@/car/service-log/presentation/tanstack/query/keys';
 import { getServiceLogsQueryOptions } from '@/car/service-log/presentation/tanstack/query/options';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
-import { queryKeySerialize } from '@/common/presentation/tanstack/query-key';
+import { queryKeySerialize } from '@/common/presentation/tanstack/query-key-serialize';
 
 export function useCostsSection() {
   const { addToast } = useToasts();
