+++ src/module/car/service-log/presentation/ui/sections/service-logs/use-service-logs.ts
@@ -5,7 +5,7 @@ import { useOwnerProfilesForCar } from '@/car/ownership/presentation/hooks/use-o
 import { queryKeys } from '@/car/service-log/presentation/tanstack/query/keys';
 import { getServiceLogsByCarIdQueryOptions } from '@/car/service-log/presentation/tanstack/query/options';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
-import { queryKeySerialize } from '@/common/presentation/tanstack/query-key';
+import { queryKeySerialize } from '@/common/presentation/tanstack/query-key-serialize';
 import { useSessionUser } from '@/user/presentation/hooks/use-session-user';
 
 interface UseServiceLogsSectionParams {
