+++ src/module/user/presentation/hooks/use-session-user.ts
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 import { useEffect } from 'react';
 
 import { useToasts } from '@/common/presentation/hook/use-toasts';
-import { queryKeySerialize } from '@/common/presentation/tanstack/query-key';
+import { queryKeySerialize } from '@/common/presentation/tanstack/query-key-serialize';
 import { getSessionUserQueryOptions } from '@/user/presentation/tanstack/query/options';
 
 export function useSessionUser() {
