+++ src/module/user/presentation/tanstack/mutation/avatar-edit.ts
