+++ src/module/user/presentation/ui/forms/avatar/use-avatar-edit.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 
 import { useToasts } from '@/common/presentation/hook/use-toasts';
-import { userAvatarEditMutationOptions } from '@/user/presentation/tanstack/mutation-options/avatar-edit';
+import { userAvatarEditMutationOptions } from '@/user/presentation/tanstack/mutation/avatar-edit';
 
 export function useUserAvatarEdit() {
   const { addToast } = useToasts();
