+++ src/module/user/presentation/ui/forms/name/use-name-edit.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 
 import { useToasts } from '@/common/presentation/hook/use-toasts';
-import { userNameEditMutationOptions } from '@/user/presentation/tanstack/mutation-options/name-edit';
+import { userNameEditMutationOptions } from '@/user/presentation/tanstack/mutation/name-edit';
 
 export function useUserNameEdit() {
   const { addToast } = useToasts();
